remove deprecated part and build (#54)
