+++ src/lexer/char-buffer/index.test.ts
@@ -39,53 +39,3 @@ describe("peekChar()", () => {
     expect(buffer.peekChar()).toEqual(expected);
   });
 });
-
-describe("pop()", () => {
-  it("pop characters", () => {
-    const buffer = new CharBuffer("ab");
-
-    expect(buffer.pop()).toBe("a");
-    expect(buffer.pop()).toBe("b");
-  });
-
-  it("pop null character if nothing to pop", () => {
-    const buffer = new CharBuffer("");
-
-    expect(buffer.pop()).toBe("\0");
-  });
-
-  it("pop null character more than once if nothing to pop", () => {
-    const buffer = new CharBuffer("");
-
-    expect(buffer.pop()).toBe("\0");
-    expect(buffer.pop()).toBe("\0");
-  });
-});
-
-describe("peek()", () => {
-  it("peek character", () => {
-    const buffer = new CharBuffer("a");
-
-    expect(buffer.peek()).toBe("a");
-  });
-
-  it("peek the same character twice", () => {
-    const buffer = new CharBuffer("a");
-
-    expect(buffer.peek()).toBe("a");
-    expect(buffer.peek()).toBe("a");
-  });
-
-  it("peek null character if nothing to pop", () => {
-    const buffer = new CharBuffer("");
-
-    expect(buffer.peek()).toBe("\0");
-  });
-
-  it("peek null character more than once if nothing to pop", () => {
-    const buffer = new CharBuffer("");
-
-    expect(buffer.peek()).toBe("\0");
-    expect(buffer.peek()).toBe("\0");
-  });
-});
